+++ RELEASE_NOTES.md
@@ -1,3 +1,36 @@
+## [0.5.4] / 20 April 2022
+- Disable `remember-entities`, but keep DData
+- Enable K8s lease
+- Run with Akka.NET v1.5.0-beta
+
+## [0.5.3] / 20 April 2022
+- Enable `remember-entities` with `state-store-mode=ddata`
+- Log the version of Akka.Cluster.Sharding being used
+- Run with Akka.NET v1.5.0-beta
+
+## [0.5.2] / 20 April 2022
+- Enable `remember-entities` with `state-store-mode=ddata`
+- Log the version of Akka.Cluster.Sharding being used
+- Run with Akka.NET v1.4.37
+
+## [0.5.1] / 20 April 2022
+- Enable `remember-entities`
+- Log the version of Akka.Cluster.Sharding being used
+- Run with Akka.NET v1.5.0-beta
+
+## [0.4.9] / 20 April 2022
+- Enable `remember-entities`
+- Log the version of Akka.Cluster.Sharding being used
+- Run with Akka.NET v1.4.37
+
+## [0.4.8] / 20 April 2022
+- Enable `remember-entities`
+- Log the version of Akka.Cluster.Sharding being used
+- Run with Akka.NET v1.4.37
+
+## [0.4.7] / 20 April 2022
+- Test with Akka.NET v1.5 with https://github.com/akkadotnet/akka.net/pull/4629 merged in
+
 ## [0.4.6] / 19 April 2022
 - Arbitrary bump for deployment purposes